Move SDK interface+mock from public pkg dir -> internal
