+++ deploy/cr.yaml
@@ -3,7 +3,7 @@ kind: PerconaServerMongoDB
 metadata:
   name: my-cluster-name
 spec:
-  #platform: openshift
+  platform: openshift
   version: "3.6"
   imagePullPolicy: Always
   secrets:
